Bug fixes after GUI overhaul and refactoring Set new GUI as main, put old as OLD_main_guizero.py
